Revert "renamed paramiko to actual used name"
This reverts commit 00fd63106847f9603f5c031903113c492c843975.
