+++ lib/ansible/plugins/connection/network_cli.py
@@ -54,7 +54,7 @@ from ansible.errors import AnsibleConnectionFailure
 from ansible.module_utils.six import BytesIO, binary_type
 from ansible.module_utils._text import to_bytes, to_text
 from ansible.plugins.loader import cliconf_loader, terminal_loader
-from ansible.plugins.connection.paramiko import Connection as _Connection
+from ansible.plugins.connection.paramiko_ssh import Connection as _Connection
 from ansible.utils.jsonrpc import Rpc
 
 try:
