+++ lib/ansible/plugins/connection/paramiko_ssh.py
